requirements.txt
cairocffi
pywayland
pywlroots
xcffib
xkbcommon
